+++ studio/components/ui/CodeEditor/CodeEditor.utils.ts
@@ -12,18 +12,19 @@ export const alignEditor = (editor: any) => {
 export const getTheme = (isDarkTheme: boolean) => {
   // [TODO] Probably need better theming for light mode
   return {
-    base: 'vs-dark', // can also be vs-dark or hc-black
+    base: isDarkTheme ? 'vs-dark' : 'vs', // can also be vs-dark or hc-black
     inherit: true, // can also be false to completely replace the builtin rules
-    colors: {
-      'editor.background': isDarkTheme ? '#1F1F1F' : '#FFFFFF',
-    },
     rules: [
-      { background: isDarkTheme ? '1F1F1F' : 'FFFFFF' },
-      { token: '', foreground: isDarkTheme ? 'D4D4D4' : '444444' },
-      { token: 'string.sql', foreground: '24B47E' },
+      { background: isDarkTheme ? '1f1f1f' : 'f0f0f0' },
+      {
+        token: '',
+        background: isDarkTheme ? '1f1f1f' : 'f0f0f0',
+        foreground: isDarkTheme ? 'd4d4d4' : '444444',
+      },
+      { token: 'string.sql', foreground: '24b47e' },
       { token: 'comment', foreground: '666666' },
       { token: 'predefined.sql', foreground: isDarkTheme ? 'D4D4D4' : '444444' },
-      // { token: '', foreground: 'ffcc00' }, // Trying to figure out how to change the border color of the row selected
     ],
+    colors: { 'editor.background': isDarkTheme ? '#1f1f1f' : '#f0f0f0' },
   }
 }
