+++ studio/hooks/misc/useStore.tsx
@@ -1,10 +1,7 @@
-import { useMonaco } from '@monaco-editor/react'
-import { useTheme } from 'common'
 import { autorun } from 'mobx'
 import { createContext, PropsWithChildren, useContext, useEffect } from 'react'
 import toast from 'react-hot-toast'
 
-import { getTheme } from 'components/ui/CodeEditor'
 import SparkBar from 'components/ui/SparkBar'
 import { IRootStore } from 'stores'
 
@@ -23,16 +20,7 @@ interface StoreProvider {
   rootStore: IRootStore
 }
 export const StoreProvider = ({ children, rootStore }: PropsWithChildren<StoreProvider>) => {
-  const monaco = useMonaco()
   const { ui } = rootStore
-  const { isDarkMode } = useTheme()
-
-  useEffect(() => {
-    if (monaco) {
-      const theme: any = getTheme(isDarkMode)
-      monaco.editor.defineTheme('supabase', theme)
-    }
-  }, [isDarkMode, monaco])
 
   useEffect(() => {
     autorun(() => {
